requirements.txt
# 依赖包列表
requests>=2.28.0
aiohttp>=3.8.0
pyyaml>=6.0
streamlit>=1.24.0
pillow>=9.0.0
python-dotenv>=0.19.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0
pydantic>=1.10.0